Obtain scientific name from taxon id
