+++ tests/processors/test_create_labware_processor.py
@@ -9,7 +9,9 @@ def test_create_labware_processor(config):
     assert CreateLabwareProcessor(schema_registry, MagicMock(), config) is not None
 
 
-def test_create_labware_processor_can_run_process(config, valid_create_labware_message, mocked_response):
+def test_create_labware_processor_can_run_process(
+    config, valid_create_labware_message, mocked_response, taxonomy_record
+):
     with patch("tol_lab_share.messages.output_feedback_message.AvroEncoderBinary") as mock_avro_encoder:
         schema_registry = MagicMock()
         publisher = MagicMock()
@@ -17,6 +19,7 @@ def test_create_labware_processor_can_run_process(config, valid_create_labware_m
         instance = CreateLabwareProcessor(schema_registry, publisher, config)
 
         with requests_mock.Mocker() as m:
+            m.get(config.EBI_TAXONOMY_URL + "/10090", json=taxonomy_record)
             m.post(config.TRACTION_URL, json=mocked_response)
             assert instance.process(valid_create_labware_message) is True
 
@@ -24,7 +27,7 @@ def test_create_labware_processor_can_run_process(config, valid_create_labware_m
         publisher.publish_message.assert_called_once()
 
 
-def test_create_labware_processor_when_traction_sends_422(config, valid_create_labware_message):
+def test_create_labware_processor_when_traction_sends_422(config, valid_create_labware_message, taxonomy_record):
     with patch("tol_lab_share.messages.output_feedback_message.AvroEncoderBinary") as mock_avro_encoder:
         schema_registry = MagicMock()
         publisher = MagicMock()
@@ -32,6 +35,7 @@ def test_create_labware_processor_when_traction_sends_422(config, valid_create_l
         instance = CreateLabwareProcessor(schema_registry, publisher, config)
 
         with requests_mock.Mocker() as m:
+            m.get(config.EBI_TAXONOMY_URL + "/10090", json=taxonomy_record)
             m.post(config.TRACTION_URL, text="This is an error", status_code=422)
 
             instance.process(valid_create_labware_message)
@@ -40,7 +44,7 @@ def test_create_labware_processor_when_traction_sends_422(config, valid_create_l
         publisher.publish_message.assert_called_once()
 
 
-def test_create_labware_processor_when_traction_sends_500(config, valid_create_labware_message):
+def test_create_labware_processor_when_traction_sends_500(config, valid_create_labware_message, taxonomy_record):
     with patch("tol_lab_share.messages.output_feedback_message.AvroEncoderBinary") as mock_avro_encoder:
         schema_registry = MagicMock()
         publisher = MagicMock()
@@ -48,6 +52,7 @@ def test_create_labware_processor_when_traction_sends_500(config, valid_create_l
         instance = CreateLabwareProcessor(schema_registry, publisher, config)
 
         with requests_mock.Mocker() as m:
+            m.get(config.EBI_TAXONOMY_URL + "/10090", json=taxonomy_record)
             m.post(config.TRACTION_URL, text="This is an error", status_code=500)
             instance.process(valid_create_labware_message)
